requirements.txt
# CUDA 12.8 

torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128

transformers==4.53.2
timm==1.0.17
pytorch-lightning==2.5.2
torchmetrics==1.7.4

pillow==11.0.0
numpy==2.1.2
scipy==1.16.0
openai
